src/face-service/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python
dlib